+++ android/app/build.gradle
@@ -154,9 +154,9 @@ android {
 }
 
 dependencies {
-    compile project(':react-native-linear-gradient')
-    compile project(':react-native-image-picker')
-    compile project(':react-native-fetch-blob')
+    implementation project(':react-native-linear-gradient')
+    implementation project(':react-native-image-picker')
+    implementation project(':react-native-fetch-blob')
     implementation fileTree(dir: "libs", include: ["*.jar"])
     implementation project(':react-native-splash-screen')
     implementation project(':react-native-fcm')
